make `colors` public
