+++ src/util/resolveConfig.js
@@ -2,7 +2,7 @@ import negateValue from './negateValue'
 import corePluginList from '../corePluginList'
 import configurePlugins from './configurePlugins'
 import defaultConfig from '../../stubs/defaultConfig.stub'
-import colors from '../../colors'
+import colors from '../public/colors'
 import log from './log'
 import { defaults } from './defaults'
 import { toPath } from './toPath'
